// src/app/shared/pipes/si-unit/si-unit.enum.ts
/**
 * represents units available for International System of Units representation
 * (also known as French Système international (d'unités) with (SI abbreviation)).
 */
export enum SiUnit {
  base = 'base',
  thousand = 'thousand',
  million = 'million',
  billion = 'billion',
  trillion = 'trillion',
}